Move namespace license tests
